ai/app/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
google-generativeai==0.3.2 